Chore: Refactor manifest verifier
